rename BlueBG & GreyBG to use full word - 'background', so these are named consistently across the codebase
