+++ src/components/Speciality/training.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Row, Col, Grid } from '../grid'
 import { SectionTitle, Subtitle, BodyPrimary } from '../Typography'
 import { Padding } from 'styled-components-spacing'
-import GreyBackground from '../Common/GreyBG'
+import GreyBackground from '../Common/GreyBackground'
 import StyledLink from '../Common/StyledLink'
 
 const TrainingStage = ({ title, body, icon }) => (
