+++ src/components/Speciality/tutorials.js
@@ -3,7 +3,7 @@ import remcalc from 'remcalc'
 import { Row, Col, Grid } from '../grid'
 import { Padding } from 'styled-components-spacing'
 import styled from 'styled-components'
-import GreyBackground from '../Common/GreyBG'
+import GreyBackground from '../Common/GreyBackground'
 import { SectionTitle, Subtitle, BodyPrimary } from '../Typography'
 import CustomisedBulletpoint from '../CustomisedBulletpoint'
 import ExternalAnchor from '../Common/ExternalAnchor'
