+++ src/components/layout.js
@@ -5,11 +5,11 @@ import { StaticQuery, graphql } from 'gatsby'
 import { Location } from '@reach/router'
 import Header from './Header'
 import './layout.css'
-import BlueBackground from './Common/BlueBG'
+import BlueBackground from './Common/BlueBackground'
 import theme from '../utils/theme'
 import GlobalStyle from '../utils/globalStyle'
 import Footer from './Footer'
-import GreyBackground from './Common/GreyBG'
+import GreyBackground from './Common/GreyBackground'
 import google from '../utils/google-json.json'
 import Cookie from './Common/CookieBanner'
