+++ src/templates/service.js
@@ -10,8 +10,8 @@ import Layout from '../components/layout'
 import CaseStudyPreview from '../components/Common/CaseStudyCards/CaseStudyPreview'
 import SeoLinks from '../components/Common/seoLinks'
 import WorkStages from '../components/Service/WorkStages'
-import GreyBackground from '../components/Common/GreyBG'
-import BlueBackground from '../components/Common/BlueBG'
+import GreyBackground from '../components/Common/GreyBackground'
+import BlueBackground from '../components/Common/BlueBackground'
 import Head from '../components/Common/Head'
 import Statement from '../components/Common/Statement'
