+++ stories/compactVideoLink.stories.js
@@ -3,8 +3,8 @@ import { storiesOf, addDecorator } from '@storybook/react'
 import Theme from './theme'
 import CompactVideoLink from '../src/components/Common/CompactVideoLink'
 import { Row } from '../src/components/grid'
-import BlueBackground from '../src/components/Common/BlueBG'
-import GreyBackground from '../src/components/Common/GreyBG'
+import BlueBackground from '../src/components/Common/BlueBackground'
+import GreyBackground from '../src/components/Common/GreyBackground'
 
 addDecorator(Theme)
