+++ stories/titleAndList.stories.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { storiesOf, addDecorator } from '@storybook/react'
 import Theme from './theme'
-import BlueBackground from '../src/components/Common/BlueBG'
-import GreyBackground from '../src/components/Common/GreyBG'
+import BlueBackground from '../src/components/Common/BlueBackground'
+import GreyBackground from '../src/components/Common/GreyBackground'
 import { Grid } from '../src/components/grid'
 import TitleAndList from '../src/components/Common/TitleAndList'
 import StyledLink from '../src/components/Common/StyledLink'
